fix(core): input injection, state hash, save/load safety
- Unified input queue with sequence numbers for deterministic ordering
- Engine-authoritative state hashing with PIC sync
- Buffer bounds checking in save/load state
- Portable little-endian serialization (V3 wire format)
- Ring buffer capacity validation (effective capacity = MAX - 1)
- V2 save state backward compatibility loader
- Fix MSVC warnings and test compatibility issues

+++ src/legends/vision_capture.cpp
@@ -214,7 +214,7 @@ void CaptureEngine::render_text_mode(Span<uint8_t> output, uint16_t width, uint1
     (void)height;  // Reserved for future font rendering
     // Text mode rendering is complex - for now, fill with a pattern
     // Full implementation would need font data
-    std::fill(output.begin(), output.end(), 0);
+    std::fill(output.begin(), output.end(), static_cast<uint8_t>(0));
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
